🔨 move GdocsConstants into gdocTypes directory
